Macro genawaiter::rc_producer

macro_rules! rc_producer {
    ($($proc_macro:tt)*) => { ... };
}

Creates a producer for use with rc::Gen.

A producer can later be turned into a generator using Gen::new.

This macro takes one argument, which should be a block containing one or more calls to yield_!.

Example

use genawaiter::{rc::Gen, rc_producer as producer, yield_};

let my_producer = producer!({
    yield_!(10);
});

let mut my_generator = Gen::new(my_producer);