genawaiter::rc

Macro gen

macro_rules! gen {
    ($body:expr) => { ... };
}

Creates a generator.

This macro takes one argument, which is the body of the generator. It should contain one or more calls to the yield_! macro.

Examples

See the module-level docs for examples.