genawaiter::sync

Attribute Macro producer_fn

#[producer_fn]

Turns a function into a producer, which can then be used to create a generator.

The body of the function should contain one or more yield_! expressions.

Examples

See the module-level docs for examples.